HELT FERDIG MED DAG 6

+++ GccApplication2/GccApplication2/Node_1/adc.h
@@ -23,6 +23,6 @@ uint8_t adc_read(uint8_t channel);
 
 int convert_horiz_JS(uint8_t A);
 
-void adc_driver(int* ADC_driver_data); //ADC_driver_data er en array som sendes som pass-by-reference
+void adc_driver(uint8_t* ADC_driver_data); //ADC_driver_data er en array som sendes som pass-by-reference
 
 void multifunction_board_test(int* ADC_driver_data);
